backend/requirements.txt
# Core dependencies from original project
firecrawl-py==1.5.0
supabase==2.10.0
openai==1.55.3
python-dotenv==1.0.1
aiohttp==3.10.10
python-dateutil==2.9.0.post0
beautifulsoup4==4.12.3
lxml==5.3.0

# FastAPI and API dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0

# Authentication removed for personal use

# Background tasks
celery==5.4.0
redis==5.2.0

# CORS support
fastapi-cors==0.0.6

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.28.0

# Development
black==24.10.0
flake8==7.1.1